fix: list page structure
