+++ pages/[[server]]/list/[list]/index/accounts.vue
